Configured website to support doxygen docstrings with Breathe extension
This should simplify the process of writing new development/design documentation for the website that describes how to use/extend the codebase. It should also reduce the burden of keeping this documentation up to date with changes made in the codebase.

A minimal working example has been included to make sure that this works properly.