Make `EnzoEFltArrayMap` copies cheaper (1/2): `StringIndRdOnlyMap`
The main change in this commit is the creation of the `StringIndRdOnlyMap` class. Using this class to implement the `str_index_map_` member of `EnzoEFltArrayMap` (instead of `std::map`) makes copies much cheaper. Copying a `std::map` instance involves cloning all of the contents of the `std::map` (which implicitly requires a heap allocation)

`StringIndRdOnlyMap` takes advantage of the fact that the contained data is just read-only and can never be mutated. With that in mind, `StringIndRdOnlyMap` stores the contents of the map in a shared pointer.
- Introducing this class facillitated some other simplifications in EnzoEFltArrayMap, that let us remove the keys_ member (these simplification were also possible with the old implementation std::map, they were just less trivial).
- We don't represent `StringIndRdOnlyMap::str_index_map_` in a `std::shared_ptr<std::map` because that would introduce an extra indirection every-time a value is looked up. There's also definitely room for some other micro-optimizations that would make this implementation faster (these are discussed in the description of the class).
- This class may also facilliate some minor optimizations in other areas of the codebase (e.g. in the `Performance` class & *possibly* in the `FieldDescr` class)

As an aside, it would probably make sense to move the implementation of `StringIndRdOnlyMap`'s constructor to src/Cello/array_StringIndRdOnlyMap.cpp, but that would cause some conflicts with the CMake PR, so I'm defering this until later.