Make `EnzoEFltArrayMap` copies cheaper (2/2): `SharedBuffer_`
This class changed the datastructure used to represent the `arrays_` member of `EnzoEFltArrayMap`. Previously it was implemented with `std::vector<T>`. Now it's implemented with the new `EnzoEFltArrayMap::SharedBuffer_<T>` template class.

This new datastructure leverages the fact that the contents of `EnzoEFltArrayMap::arrays_` are never mutated* to make copies much cheaper.
- This is accomplished by having a `std::shared_ptr` manage the lifetime of the container's allocation. Thus, when `EnzoEFltArrayMap::arrays_` is copied, the shared_ptr is copied instead of requiring the entire allocation to be cloned.
- As a point of clarification: The elements stored within `arrays_` are `CelloArray`s. `EnzoEFltArrayMap` does not permit any mutations to `CelloArray`s. But, since `CelloArray` is a type of smart pointer, users are still able to mutate the array data associated with a `CelloArray`.
- We don't use `std::shared_ptr<std::vector<T>>` because that would introduce extra indirections for every access

With this change and the introduction of `StringIndRdOnlyMap` (in the previous commit), the act of copying `EnzoEFltArrayMap` no longer requires any heap allocations. The amount data that is copied is also significantly reduced. The tradeoff is incrementing the reference counts for 2 `std::shared_ptr`s.

As an aside: our exact approach for implementing `EnzoEFltArrayMap::SharedBuffer_<T>` was selected in order to minimize conflicts with other changes to `EnzoEFltArrayMap` introduced in PR #145 (while merging both changesets, the `EnzoEFltArrayMap::SharedBuffer_<T>` functionallity will be transferred to the `VecBackedCArrCollec_` class)