Fixed bug related to adding a nonexistent field to a field group
This bug arose because of an annoying property of `std::map`:
- In this scenario, `field_index` is an instance of `std::map<std::string,int>` that maps field names to field indices.
- While adding fields to groups, based on an input configuration file, the old code checked for the index associated with a field name, `field`, using `field_index[field]`.
- Unfortunately, if `field` is not already contained by `field_index`, a new key-value pair is inserted into `field_index` (This is required to support the syntax like `field_index["density"] = 3;`).
- As a result, when users tried to add previously undeclared fields to a group, a different field was added to the group instead.

The simple way to resolve this would be to replace `field_index[field]` with `field_index.at(field)`. In cases where `field_index` does not already contain field, an exception would be raised.

A different solution used `std::map::find` was adopted to provide users with more meaningful error messages.