"accretion" methods now have an additional parameter: "ang_mom_threshold_radius_cells", which specifies the angular momentum threshold radius in terms of the minimum cell width (across all dimensions) at the highest refinement level.
If "conserve_angular_momentum" is true, the the angular momentum of the gas in a given cell is conserved
during accretion if and only if the distance of the cell from the sink particle is greater than the
angular momentum threshold radius.

If "conserve_angular_momentum" is false, the "ang_mom_threshold_radius_cells" parameter is ignored.