The ".libconfig" and ".parameters" output files now always contain a "Particle:mass_is_mass" parameter, which will nearly always have the value "dummy", but there is a small chance it will have a different value.
The purpose of this parameter is to be a flag for the Enzo-E frontend of yt, indicating that particle masses
should be treated as quantities with dimensions of mass, rather than quantities with dimensions of density. It
is only important that this parameter exists, it does not matter what value it has.