Reorganized the Circularly polarized alfven wave.
While I was here, I decided to refactor the circularly polarized alfven wave so that it used lambda functions to define its vector initializers. Doing this makes the code a lot easier to read - all of the code specific to initializing a Circularly-Polarized Alfven Wave is now located in a single spot (previously it was split into 2 different places).

I would like to clean this all up a little more, but it's not a great use of my time