Address PR comments and minor bugfix in EnzoInitialGrackleTest
This comment addresses the following 3 PR comments:
1. We now comment out more fields required by Grackle to further test automatic field creation.

2. Fixed an issue with the length units definition. We had a comment stating that we were defining the length unit to be 1 pc, but were actually setting the length unit to 0.01 pc. I choose not to actually change the length unit definition (as was recommended by @gregbryan) since that would require us to update our reference answers. Instead, I updated the comment.

3. Better documented why we use such long timesteps (i.e. for increasing the chance of exposing SMP bugs)

While modifying input/Grackle/grackle.incl, I deleted the "gamma" field. This was a holdover from a much earlier version of EnzoMethodGrackle. The current implementation does not support this field. This revealed that EnzoInitialGrackleTest was still accessing this field (it wasn't doing anything of substance with it). Therefore, I deleted the mentions of the "gamma" field from EnzoInitialGrackleTest (we can always add it back in the future).