Addressed minor oversight in the ported tests
In the process of porting the tests that compared simulation times of Enzo-E simulations (and the C++ unit tests) a minor oversight occured that reduced the functionallity of these tests. The new system assumed that the exit code of the C++ always indicated whether the test passed or not. However, that is not the case (as long as the program gracefully exits, it reports an exit code of 0).

As a consequence, a test could fail (e.g. the final simulation time could be different from the expected final time), but CTest would still report the test as passing. As a concrete example, one can:
- modify line 79 of `input/PPM/ppm.incl` so that you replace `time_final  = [1.04069762624371];` with something like `time_final = [10.0];`.
- then when you execute `ctest --verbose -R PPM-1` from the build directory, `ctest` reports that the test failed, even though the output clearly states the final time of the simulation is incorrect.

My first instinct to solve this was to pipeline running the simulation, dumping stdout to a log file, and then call a script that parse the logfile and exits with the appropriate exit code. Unfortunately, the way that ctest works requires that a test is entirely encapsulated by a single command (that takes optional potential arguments).

To that end, I have introduced a python script called `tools/run_cpp_test.py` to do all 3 steps for us at once. The script wraps and executes an arbitrary command (much like a profiling tool like `time` or `perf-stat`) and then parses the standard output stream to check for indications that the test may have failed (I have ported this logic from the `build.sh` script from just before we transitioned to cmake).

As a concrete example, one could execute the following command from the repository's root directory:
```
python3 tools/run_cpp_test.py --output-dump sim_output.log path/to/enzo-e input/PPM/method_ppm-1.in
```
Note that the optional `--output-dump` argument specifies where the script writes a copy of stdout (It still continues to write a copy of stdout to stdout). This is what the script parses to check if the program succeeded. If you don't specify this, the program will write a copy of stdout to a temporary file that will be deleted once the program ends.

There are 2 other alternative approaches for addressing this problem:
1. We continue to use the wrapper script approach, but rewrite it in a different language (like bash or c++)
2. We could modify the unit testing machinery to always exit with a non-zero exit code if it detects a testing error. Likewise, we would need to modify Cello/Enzo-E to exit with a non-zero exit code if the testing machinery detects an unexpected completion time/completion cycle.

I actually really like the 2nd approach as a longer term solution. But, that will take some additional work and the current approach fixes things in the meantime.