Refactored CArrCollec to make use of the SharedBuffer_ machinery from EnzoEFltArrayMap when storing CelloArrays as a single contiguous 4D CelloArray
This change was required to preserve fast copies when EnzoEFltArrayMap tracks CelloArrays as an array of pointers after the previous merge commit. I kept this as a separate commit in case I broke something unintentionally.

As part of this change I renamed the class detail::VecBackedCArrCollec_ to now be called detail::ArrOfPtrsCArrCollec_ which is a more accurate name.