Make incremental local rebuilds of the website faster
Prior to this commit, sphinx would rerun doxygen on the source directory everytime you tried to rebuild the documentation locally. The only way to avoid this was creating an environment variable called `SKIPDOXYGEN` and assigning it a value of `true`.

As a solution to this problem, we now compute an md5 checksum for the entire src and doxygen directories and cache the values in 2 text files. If both values remain unchanged, we now skip the rebuild. With this change, this cuts the time for re-executing sphinx (without any modifications from 47.3 seconds down to 0.8 seconds.

Another (potentially better) solution is to check the last modification time of the directories.