Simplified mechanism for avoiding doxygen during documentation rebuilds
Previously we used md5 checksums to determine whether we should rebuild doxygen documentation (during documentation rebuilds). Now we use cached mtimes (stored by the file system). The logic is simpler and this is slightly faster (by about ~0.2 s).

Note this rudimentary speed test was carried out with a SSD. On a system with a mechanical HD, the speed difference may be more significant (since we no longer need to read the contents of every file)

The speed difference may be larger on systems that use