Updated expected test answers for Grackle in single-precision mode
These test answers changed quite a few commits ago when I converted EnzoMethodGrackle::ResetEnergies and EnzoInitialGrackleTest::enforce_block to load the value of gamma from EnzoPhysicsFluidProps. As part of this change, the precision of gamma changed from double to enzo_float in these methods (which is the primary reason for needing to update the test answers).

There's a chance we may want to stick with double precision, but I'm going to punt that question