Removed all mentions of "gamma" field from input files.
The "gamma" field first appeared in the input files during an earlier version of Enzo-E. However back in PR #1, it was removed (Grackle no longer uses it). If we wanted, it would be straightforward to output it as a derived quantity, but for the short-term, it's better not to include it (to avoid confusing new users).

While I was doing this, I fixed a couple of cases where the "temperature", "pressure", "cooling_time" fields were all listed as part of the "color" field. This also crept in as part of an older version of a Grackle input file.