fix charm++ SMP detection - take 2
Overview
--------

In PR #256 I attempted to fix a bug in the FindCharm CMake module that arises while checking if charm++ was built in SMP mode for newer versions of charm++.

I made a minor oversight and those changes caused CMake to always conclude that (newer versions of) charm++ always use SMP mode.

I must have forgetten to rebuild Enzo-E from scratch using non-SMP builds of charm++. Sorry about that.

Problem Descriptions
--------------------

In older versions of charm++, the CMK_SMP macro was only defined in SMP builds. Thus, we used CMake's CHECK_SYMBOL_EXISTS macro to check whether the macro was defined. As an aside, when the macro is defined, it has a value of 1.

In newer versions of charm++, the CMK_SMP is now defined in a different header. Additionaly, the macro is now ALWAYS defined, and the value is set as follows:
- 1 when using SMP mode (this was the case previously)
- 0 when not using SMP mode

My changes in PR #256 just checked if CMK_SMP was defined.

Note: the version cutoff seems to occur around version 7.0 of charm++, so we probably don't need to worry about supporting the check for the older version (but I kept it for now).

Solution
--------

CMake now queries the existence and value of the CMK_SMP (via a test program) to determine whether charm++ was built in SMP mode.

I have also confirmed that CMake draws the correct conclusion for both SMP and non-SMP versions of CMake.