Remove hard python dependency during compilation
Since the transition to CMake, we have required python3 as a hard dependency.
  * This choice was initially made so that we could check selectively enable/disable some tests based on whether certain python module could be found.
  * This choice was somewhat controversial because it substantially complicates the process of installing Enzo-E.
  * This makes first-time installation very frustrating since the quickstart guide doesn't address any of this.
  * A change I made shortly after the introduction of CMake (to fix a bunch of tests that didn't properly report their results) probably exacerbated things.

This changeset introduces a new compromise.
  * CMake no longer looks for python3 at all (we implicitly assume that all systems that want to run tests have a python3 interpretter)
  * CTest now uses the "active" python3 interpretter to run tests.

However, we still perform a check for whether the python3 interpretter has the required modules test for the yt-related modules. We do this as part of a ctest "fixture". If the modules can't be found, the dependent tests are skipped (CTest loudly anounces such tests are not run because of `Failed test dependencies: python_has_yt_and_other_modules`

Of course, this will all be less of a problem once we migrate the yt-based test to the answer testing framework.