Fixed the failing docs-build circleci test
In detail, this test tries to build the website documentation and fails if any warnings get raised. The test recently started failing (e.g. in PR #280 and PR #263) because the ``sphinx`` developers introduced a new warning to indicate that a configuration option was specified in an outdated format.

In more detail:

  * I believe that when ``sphinx`` originally generated our documentation configuration file, ``doc/source/conf.py``, the ``intersphinx`` extension was enabled by default (for the sake of illustrating how to use extensions).

  * One of the configuration options that was autogenerated for that extension, called ``intersphinx_mapping``, was specified in a format that was used before version 1.0 of ``sphinx``

  * In a recent update to ``sphinx``, the developers introduced a deprecation warning indicating that the format of that parameter is deprecated (and will become invalid in an upcoming release)

To resolve this issue, I disabled the ``intersphinx`` extension since we don't use it at all (to avoid issues like this in the future).