Add support for implicit casting of ViewMap<T> to ViewMap<const T>
This is the major "fix" that is needed to address GH Issue #172.

We removed the old behavior, where we assumed that a const ViewMap represented readonly data. There was a soundness problem issue with this behavior. To illustrate this problem, let's consider the following example:

One of the primary reasons one might use `const` in the context of a ViewMap is in the context of a function argument. Consider the following hypothetical function declaration:

```
CelloView<enzo_float, 3> calculate_pressure(ViewMap<enzo_float> arg,
                                            double density_floor);
```

At a glance, one might gather that this function is intended for computing pressure, and applying the density floor. However, it's a little ambiguous what the role of `density_floor` is in this context. For example, is the "density" view getting updated by `density_floor`?

Assuming that the original desired behavior is sound, the author might instead write the following function signature in order to signal that `density_floor` only affects the value of the ouput View, and that none of the elments in any view held by `arg` are mutated.

```
CelloView<enzo_float, 3> calculate_pressure(const ViewMap<enzo_float> arg,
                                            double density_floor);
```

Unfortunately, the soundness problem (described in GH Issue #172) lets you write code like the following:
```
CelloView<enzo_float, 3> calculate_pressure(const ViewMap<enzo_float> arg,
                                            double density_floor)
{
  ViewMap<enzo_float> tmp(arg);
  // you are now free to mutate tmp["density"]

  // ...
}
```

The only way to fix this while maintaining the existing semantics is to remove the ability to make copies or to move instances of `ViewMap`. This seems too limiting.

Instead, I have changed the behavior so that it is more similar to what is done for for CelloView. Now, we support implicit conversions from ViewMap<T> to ViewMap<const T>. If a developer wants to indicate that an argument is not mutated they would specify:
```
CelloView<enzo_float, 3> calculate_pressure(ViewMap<const enzo_float> arg,
                                            double density_floor);
```