Moved EnzoInitialPm from obsolete to initial/obsolete
This class was defined in the obsolete subdirectory. It was my
originally my understanding that everything in that directory could be
deleted.
-> Thus I initially deleted the file
-> While the class is used in a handful of input files, none of these
   files were used by the automated testing infrastructure. Thus, these
   were deleted too.

While reviewing the PR James indicated that these files should be kept.
Consequently I editted this commit and restored the EnzoInitialPm class
and it's associated input files (other than the input files in
`input/Obsolete`). Since it's my goal to delete `src/Enzo/obsolete` I
instead opted to move the code declaring/defining this class into
`src/Enzo/initial/obsolete`