Moved EnzoBlock::Solve{Hydro,MHD}Equations member-functions to static methods of EnzoMethodPpm and EnzoMethodPpml.
These functions were the convention for integrators to be encapsulated by Method Objects was (presumably) not strongly established. This is rectifying that situation.

In the future, it probably makes sense to consolidate
- the source file with the definition for EnzoMethodPpm::SolveHydroEquations with the source file defining the rest of EnzoMethodPpm
- the source file with the definition for EnzoMethodPpml::SolveMHDEquations with the source file defining the rest of EnzoMethodPpml

However, that consolidation step has been left to a separate PR from this changeset in order to make merge conflicts easier to address

I have also deleted the declaration of EnzoBlock::SolveHydroEquations3 because it doesn't correspond to any function