Shed unnecessary PUP routines from hydro-infrastructure.
Changes introduced in PR #315 make it unnecessary for individual classes in the hydro-infrastructure to have PUP routines. Instead of directly serializing the individual classes, the changes introduced in that PR direct the solver to serialize the parameters used to configure the solver, and then reconstructs all of the component classes after a restart. This simplifies things quite a bit.

This commit sheds the now-unnecessary PUP routines.