Fix GH Issue #357 (#358)
The problem in that issue was relatively straight-forward. A while back
we introduced a new way to initialize the global EOS object. At that
time, I deprecated the `Field:gamma` parameter, but for the sake of
backwards compatability I used `Field:gamma` to initialize an ideal EOS.

The error reported in Issue #357 occurred because `Field:gamma` was
assigned a value of `1.0` to reflect the fact the PPML solver uses an
isothermal function. When the code tried to use this to initialize
`EnzoEOSIdeal`, the constructor complained because an adiabatic index of
1 is invalid for the current implementation.

To resolve this situation, I added a more informative error message
explaining how the user should fix the problem. I also fixed the PPML
test problem's input file to reflect the "correct" way to handle this
case.

Along the way, I uncovered a minor bug in the way the `Parameters` class
handles parsing a string with a fallback value (where the fallback value
is specified as a string-literal). Essentially, we were missing an
overload telling the function how to properly handle this case.
Previously when the compiler saw a string literal passed, it coerced the
pointer to a boolean (it asked if the pointer was NULL or not). I have
now fixed that case.