EnzoUnits use override specifier and Units bugfix
This commit tweaks the declaration of the virtual methods in
`EnzoUnits` to make use of the `override` specifier.

- the `override` specifier signals (to the reader and to the compiler)
  that the labelled member-function is intended to override/replace
  an existing virtual member-function from the base-class.
- if the compiler can't find an existing virtual member-function to be
  overriden, it will raise a compiler error.

In the process of doing this, I found a bug. I realized that `EnzoUnits`
was trying to override the `density` member-function even though it is
not declared as virtual function of the `Units` class.
- prior to this commit, if you called `units->density()`, you would get a
  different answer based on whether `units` was declared as `Units*` or as
  `EnzoUnits*`.
- This commit resolves this bug by declaring the `Units::density` method
  as a virtual function