Removed sprintf usage within parse.y
All `sprintf` were replaced with snprintf (this was prompted by the
fact that apple seems to have attached a depraction attribute to the
`sprintf` function in the version of libc that they ship.

The net result is a little "safer" (before we just assumed that the
buffer was big enough to hold the generated expression. Now, we
truncate at the end of the buffer).

After modifying parse.y, I regenerated the parse.tab.c and parse.tab.h
files from parse.y