Removed unnecessary Assertion from EnzoInitialGrackleTest::enforce_block
This check was trying to see whether block was a NULL pointer. A huge
guarantees ensures that Block can NEVER be a NULL pointer.

This is maybe the only Initializer where we perform this check and its
really just noise. We don't want to encourage people to copy this check