Suppress inconsistent ``override`` warnings.
There are a few spots in the codebase where we have started to make use
of the ``override`` keyword (it tells the compiler to make sure that you
are actually overriding an existing virtual function - this is an
extremely useful feature).

The clang compiler provides warnings if some virtual methods within a
given class are annotated with ``override`` while other methods are not
annotated. (It dislikes the inconsistency -- and I think it's a fair
complaint)

This change addresses that warning by using ``override`` more
consistently throughout the class (where we had already used it once it
in the past)