Moved `add_subdirectory(src/External)` in CMakeLists.txt
It was moved to earlier in the file. The idea is to insolate compilation
of "vendored dependencies" from the arbitrary flags that users specify
for building Enzo-E/Cello.

In practice, this doesn't matter right now. But in the future, it could
matter a lot more! For example a dependency could entirely break when,
-> using unsafe-floating point optimizations...
-> using compiler-flags to properly handle Enzo-E's fortran dialect